requirements.txt
qq-botpy
pyyaml
websocket_client